fix(stackable-operator): Re-export versioned error types
