+++ crates/stackable-operator/src/crd/s3/bucket/v1alpha1_impl.rs
@@ -1,14 +1,26 @@
 //! v1alpha1 specific implementations for S3 buckets.
 
-use snafu::ResultExt as _;
+use snafu::{ResultExt as _, Snafu};
 
 use crate::{
     client::Client,
-    crd::s3::bucket::{
-        BucketError, ResolveConnectionSnafu, RetrieveS3ConnectionSnafu,
-        v1alpha1::{InlineBucketOrReference, ResolvedBucket, S3Bucket},
+    crd::s3::{
+        bucket::v1alpha1::{InlineBucketOrReference, ResolvedBucket, S3Bucket},
+        connection::v1alpha1::ConnectionError,
     },
 };
+
+#[derive(Debug, Snafu)]
+pub enum BucketError {
+    #[snafu(display("failed to retrieve S3 connection '{s3_connection}'"))]
+    RetrieveS3Connection {
+        source: crate::client::Error,
+        s3_connection: String,
+    },
+
+    #[snafu(display("failed to resolve S3 connection"))]
+    ResolveConnection { source: ConnectionError },
+}
 
 impl InlineBucketOrReference {
     pub async fn resolve(
